requirements.txt
pandas
matplotlib.pyplot
PIL
wordcloud